some progress on ga

+++ src/deltaKinematics.h
@@ -61,7 +61,7 @@ public:
     
     void drawWorkingPointCloud();
     
-    int calculatePointCloudSize(int, int); //this is the fitness function that the ga runs
+    int calculatePointCloudSize(float, float); //this is the fitness function that the ga runs
 };
